Aqui sigo con las pruebas Unitarias (Andres)

+++ src/main/java/com/spring/proyecto2/dao/AgendaDao.java
@@ -18,6 +18,6 @@ import com.spring.proyecto2.model.Persona;
  */
 
 @Repository
-public interface AgendaDao extends JpaRepository<Persona, Long> {
+public interface AgendaDao extends JpaRepository<Persona, Integer> {
 
 }
